Debug

+++ datasets/IemocapDataset.py
@@ -140,7 +140,7 @@ class IemocapDataset(object):
         end = self.df.loc[idx, 'end']
         frames = sample_frames(video_name, num_frames=4, start=start, end=end)
         frames = [process_image(Image.fromarray(frame)) for frame in frames]
-        frames = torch.cat(frames).to(torch.float16)
+        frames = torch.cat(frames)
         transcription = self.df.loc[idx, 'transcription']
 
         sample = {
